requirements.txt
aiogram
mysql-connector-python
pillow
pillow_heif
rembg
onnxruntime
logging